requirements.txt
Flask
PyMuPDF
pypdf
pdfminer.six
openai
pinecone-client
python-dotenv
